Move client back to controller
