+++ controller/api/public/client.go
@@ -1,9 +1,11 @@
-package conduit
+package public
 
 import (
 	"bufio"
 	"bytes"
+	"encoding/binary"
 	"fmt"
+	"io"
 	"net/http"
 	"net/url"
 
@@ -15,6 +17,15 @@ import (
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
+)
+
+const (
+	ApiRoot             = "/" // Must be absolute (with a leading slash).
+	ApiVersion          = "v1"
+	JsonContentType     = "application/json"
+	ApiPrefix           = "api/" + ApiVersion + "/" // Must be relative (without a leading slash).
+	ProtobufContentType = "application/octet-stream"
+	ErrorHeader         = "conduit-error"
 )
 
 type client struct {
@@ -143,4 +154,38 @@ func newClient(apiURL *url.URL, httpClientToUse *http.Client) (pb.ApiClient, err
 		serverURL: apiURL.ResolveReference(&url.URL{Path: ApiPrefix}),
 		client:    httpClientToUse,
 	}, nil
+}
+
+func fromByteStreamToProtocolBuffers(byteStreamContainingMessage *bufio.Reader, errorMessageReturnedAsMetadata string, out proto.Message) error {
+	//TODO: why the magic number 4?
+	byteSize := make([]byte, 4)
+
+	//TODO: why is this necessary?
+	_, err := byteStreamContainingMessage.Read(byteSize)
+	if err != nil {
+		return fmt.Errorf("error reading byte stream header: %v", err)
+	}
+
+	size := binary.LittleEndian.Uint32(byteSize)
+	bytes := make([]byte, size)
+	_, err = io.ReadFull(byteStreamContainingMessage, bytes)
+	if err != nil {
+		return fmt.Errorf("error reading byte stream content: %v", err)
+	}
+
+	if errorMessageReturnedAsMetadata != "" {
+		var apiError pb.ApiError
+		err = proto.Unmarshal(bytes, &apiError)
+		if err != nil {
+			return fmt.Errorf("error unmarshalling error from byte stream: %v", err)
+		}
+		return fmt.Errorf("%s: %s", errorMessageReturnedAsMetadata, apiError.Error)
+	}
+
+	err = proto.Unmarshal(bytes, out)
+	if err != nil {
+		return fmt.Errorf("error unmarshalling bytes: %v", err)
+	} else {
+		return nil
+	}
 }
