--- pkg/conduit/conduit.go
@@ -1,10 +1,0 @@
-package conduit
-
-const (
-	ApiRoot             = "/" // Must be absolute (with a leading slash).
-	ApiVersion          = "v1"
-	JsonContentType     = "application/json"
-	ApiPrefix           = "api/" + ApiVersion + "/" // Must be relative (without a leading slash).
-	ProtobufContentType = "application/octet-stream"
-	ErrorHeader         = "conduit-error"
-)
